Protect shift and shift_dimension routes, and make then team specific.

+++ backend/src/routes/shift_dimension_routes.py
@@ -2,48 +2,76 @@ from dataclasses import asdict
 from typing import Dict, List
 
 import humps
-from fastapi import APIRouter, Body, HTTPException
+from fastapi import APIRouter, Depends, HTTPException
 from pydantic import TypeAdapter
 
 from core.shift import ShiftDimension, ShiftProperty
 from routes.api_model import NewShiftDimensionMessage, ShiftDimensionMessage
 from routes.shift_routes import shift_property_to_api_msg
 from scripts.setup_database import shift_db, shift_dimension_db, shift_property_db
+from services.authentication.authn_services import authn_verify_session
+from services.authentication.authn_types import SessionContainerType
+from services.authorization.authz_services import permit_check
 
 router = APIRouter()
 
 
-@router.post("/shift-dimensions")
-def create_shift_dimension(
-    name: str = Body(...),
-    entry_type: str = Body(..., alias="entryType"),
-    entry_options: List[str] = Body(..., alias="entryOptions"),
+@router.post("/shift-dimensions/teams/{team_id}")
+async def create_shift_dimension(
+    team_id: str,
+    shift_dimension: ShiftDimensionMessage,
+    session: SessionContainerType = Depends(authn_verify_session()),
 ) -> NewShiftDimensionMessage:
-    shift_dimension = shift_dimension_db.create_shift_dimension(
-        name, entry_type, entry_options
-    )
+    if not await permit_check(
+        session.get_user_id(), "create-shift-dimension", "team", team_id
+    ):
+        raise HTTPException(
+            status_code=403,
+            detail="You do not have permission to create a shift dimension",
+        )
+    sd_data = api_msg_to_shift_dimension(shift_dimension)
+    sd_created = shift_dimension_db.create_shift_dimension(sd_data)
     properties = []
-    if entry_type == "bool":
-        shifts = shift_db.get_shifts()
+    if sd_created.entry_type == "bool":
+        shifts = shift_db.get_shifts(team_id)
         for shift in shifts:
             properties.append(
-                shift_property_db.create_shift_property(shift, shift_dimension, False)
+                shift_property_db.create_shift_property(shift, sd_created, False)
             )
-    return new_shift_dimension_to_api_msg(shift_dimension, properties)
+    return new_shift_dimension_to_api_msg(sd_created, properties)
 
 
-@router.get("/shift-dimensions")
-def get_shift_dimensions() -> List[ShiftDimensionMessage]:
-    shift_dimensions = shift_dimension_db.get_shift_dimensions()
+@router.get("/shift-dimensions/teams/{team_id}")
+async def get_shift_dimensions(
+    team_id: str,
+    session: SessionContainerType = Depends(authn_verify_session()),
+) -> List[ShiftDimensionMessage]:
+    if not await permit_check(
+        session.get_user_id(), "read-shift-dimensions", "team", team_id
+    ):
+        raise HTTPException(
+            status_code=403,
+            detail="You do not have permission to read shift dimensions",
+        )
+    shift_dimensions = shift_dimension_db.get_shift_dimensions(team_id)
     return [shift_dimension_to_api_msg(sd) for sd in shift_dimensions]
 
 
-@router.put("/shift-dimensions/{shift_dimension_id}")
-def update_shift_dimension(
+@router.put("/shift-dimensions/{shift_dimension_id}/teams/{team_id}")
+async def update_shift_dimension(
     # pylint: disable=W0613
     shift_dimension_id: str,
+    team_id: str,
     shift_dimension: ShiftDimensionMessage,
+    session: SessionContainerType = Depends(authn_verify_session()),
 ) -> ShiftDimensionMessage:
+    if not await permit_check(
+        session.get_user_id(), "update-shift-dimension", "team", team_id
+    ):
+        raise HTTPException(
+            status_code=403,
+            detail="You do not have permission to update a shift dimension",
+        )
     existing_shift_dim = shift_dimension_db.get_shift_dimension_by_id(
         shift_dimension_id
     )
@@ -56,8 +84,19 @@ def update_shift_dimension(
     return shift_dimension_to_api_msg(updated_shift_dimension)
 
 
-@router.delete("/shift-dimensions/{shift_dimension_id}")
-def delete_shift_dimension(shift_dimension_id: str) -> Dict:
+@router.delete("/shift-dimensions/{shift_dimension_id}/teams/{team_id}")
+async def delete_shift_dimension(
+    shift_dimension_id: str,
+    team_id: str,
+    session: SessionContainerType = Depends(authn_verify_session()),
+) -> Dict:
+    if not await permit_check(
+        session.get_user_id(), "delete-shift-dimension", "team", team_id
+    ):
+        raise HTTPException(
+            status_code=403,
+            detail="You do not have permission to delete a shift dimension",
+        )
     shift_property_db.delete_shift_properties_by_shift_dimension_id(shift_dimension_id)
     shift_dimension_db.delete_shift_dimension(shift_dimension_id)
     return {"message": "shift deleted"}
